--- fluentai/services/card_gen/tts/test.py
@@ -1,3 +1,0 @@
-import gtts
-
-print(gtts.lang.tts_langs())
